change average property to type.FLOAT for exact averages

+++ src/trueToSize/trueToSize.model.js
@@ -9,7 +9,7 @@ module.exports = (sequelize, type) => sequelize.define('trueToSize', {
     defaultValue: [],
   },
   average: {
-    type: type.INTEGER,
+    type: type.FLOAT,
     defaultValue: 0,
   },
 });
